+++ pkg/repositories/errors.go
@@ -5,3 +5,7 @@ import "github.com/giantswarm/microerror"
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
 }
+
+var repositoryNotFoundError = &microerror.Error{
+	Kind: "repositoryNotFoundError",
+}